Move CompileException class into a separate file
