+++ src/SIL.Machine.Morphology.HermitCrab/AnalysisStratumRule.cs
@@ -99,17 +99,6 @@ namespace SIL.Machine.Morphology.HermitCrab
             }
         }
 
-        public class CompileException : Exception
-        {
-            public CompileException() { }
-
-            public CompileException(string message)
-                : base(message) { }
-
-            public CompileException(string message, Exception inner)
-                : base(message, inner) { }
-        }
-
         public IEnumerable<Word> Apply(Word input)
         {
             if (_morpher.TraceManager.IsTracing)
